\name{foreYield-package}
\alias{foreYield-package}
\alias{foreYield}
\docType{package}
\title{
\packageTitle{foreYield}
}
\description{
\packageDescription{foreYield}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{foreYield}
\packageIndices{foreYield}
}
\author{
\packageAuthor{foreYield}

Maintainer: \packageMaintainer{foreYield}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
