\name{foreYield-package}
\alias{foreYield-package}
\alias{foreYield}
\docType{package}
\title{foreYield
\packageTitle{foreYield}
}
\description{
\packageDescription{foreYield}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{foreYield}
\packageIndices{foreYield}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{foreYield}

Maintainer: \packageMaintainer{foreYield}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
#This should allow you to just follow the instructions on screen
virgilio()
}
